fix: 將 set 的重量改為整數型態

+++ models/workout_set.py
@@ -16,6 +16,6 @@ class WorkoutSet(Base):
     workout_id = Column(Integer, ForeignKey('workouts.id'))
     exercise_id = Column(Integer, ForeignKey('exercises.id'))
     reps = Column(Integer)
-    weight = Column(DECIMAL)
+    weight = Column(Integer)
     workout = relationship("Workout")
     exercise = relationship("Exercise")
